# api/public/api/openapi.json
{"openapi":"3.0.0","info":{"title":"Stellario - API","version":"1.0.0","description":"Doc de sp\u00e9cification pour API.","termsOfService":"","license":null,"contact":[],"x-logo":{"url":"https:\/\/cdn-icons-png.flaticon.com\/512\/599\/599505.png","altText":"Stellario - API","backgroundColor":""}},"components":{"securitySchemes":[],"schemas":[]},"servers":[{"url":"http:\/\/localhost:8000","description":"Environnement de stagging."}],"paths":{"\/api\/idoc":{"get":{"tags":["general"],"operationId":"idoc","description":"","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":null}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/idoc\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/idoc\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/idoc\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/v1\/stars":{"get":{"tags":["Stars"],"operationId":"Get stars","description":"```StarController::class```","parameters":[{"in":"query","name":"search","description":"Input search, can be first name, last name, both or popularity.","required":false,"schema":{"type":"string","example":"dany bitard"}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/v1\/stars\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/v1\/stars\");\n\nlet params = {\n    \"search\": \"dany bitard\",\n};\n\nObject.keys(params).forEach(key => url.searchParams.append(key, params[key]));\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$query = [\n    \"search\" => \"dany bitard\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/v1\/stars?\" . http_build_query($data));\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"post":{"tags":["Stars"],"operationId":"Store star","description":"```StarController::class```","requestBody":{"description":"```StarController::class```","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["first_name","last_name","popularity","description"],"properties":{"first_name":{"type":"string","example":"Dany","description":"First name of the star."},"last_name":{"type":"string","example":"Bitard","description":"Last name of the star."},"popularity":{"type":"integer","example":69,"description":"Popularity of the star ```1 to 99```."},"description":{"type":"string","example":"","description":"Description of the star."}},"example":{"first_name":"Dany","last_name":"Bitard","popularity":69,"description":""}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"La description est requise.","errors":{"description":["La description est requise."]}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/v1\/stars\" \n    -d \"first_name\"=\"Dany\" \\\n    -d \"last_name\"=\"Bitard\" \\\n    -d \"popularity\"=\"69\" \\\n    -d \"description\"=\"\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/v1\/stars\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"first_name\": \"Dany\",\n    \"last_name\": \"Bitard\",\n    \"popularity\": \"69\",\n    \"description\": \"\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"first_name\" =>  \"Dany\",\n    \"last_name\" =>  \"Bitard\",\n    \"popularity\" =>  \"69\",\n    \"description\" =>  \"\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/v1\/stars\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/v1\/stars\/{star}":{"get":{"tags":["Stars"],"operationId":"Get star","description":"```StarController::class```","parameters":[{"in":"path","name":"star","description":"ID of the star.","required":true,"schema":{"type":"integer","example":1}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"data":{"id":1,"first_name":"Carroll","last_name":"Kautzer","face":{"origin":"public\/face-1.png","path":"http:\/\/localhost:8000\/storage\/face-1.png","mime":"image\/png","size":1193892},"description":"Debitis ipsum consequatur quaerat recusandae aperiam harum. Nisi qui qui qui est placeat. Ad et impedit delectus itaque. Officia ipsam et quibusdam placeat quia vero. Optio sapiente velit voluptas est odio. Ratione dolor sapiente aut occaecati quia.","popularity":93,"updated_at":"25\/07\/2023 \u00e0 09:06"}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"put":{"tags":["Stars"],"operationId":"Update star","description":"```StarController::class```","requestBody":{"description":"```StarController::class```","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["first_name","last_name","popularity","description"],"properties":{"first_name":{"type":"string","example":"Dany","description":"First name of the star."},"last_name":{"type":"string","example":"Bitard","description":"Last name of the star."},"popularity":{"type":"integer","example":69,"description":"Popularity of the star ```1 to 99```."},"description":{"type":"string","example":"","description":"Description of the star."}},"example":{"first_name":"Dany","last_name":"Bitard","popularity":69,"description":""}}}}},"parameters":[{"in":"path","name":"star","description":"ID of the star.","required":true,"schema":{"type":"integer","example":1}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"La description est requies.","errors":{"description":["La description est requies."]}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PUT \"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\" \n    -d \"first_name\"=\"Dany\" \\\n    -d \"last_name\"=\"Bitard\" \\\n    -d \"popularity\"=\"69\" \\\n    -d \"description\"=\"\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"first_name\": \"Dany\",\n    \"last_name\": \"Bitard\",\n    \"popularity\": \"69\",\n    \"description\": \"\",\n})\n\nfetch(url, {\n    method: \"PUT\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"first_name\" =>  \"Dany\",\n    \"last_name\" =>  \"Bitard\",\n    \"popularity\" =>  \"69\",\n    \"description\" =>  \"\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"delete":{"tags":["Stars"],"operationId":"Remove star","description":"```StarController::class```","parameters":[{"in":"path","name":"star","description":"ID of the star.","required":true,"schema":{"type":"integer","example":1}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"toast":{"message":"Star supprim\u00e9e aves succ\u00e8s","type":"success"}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X DELETE \"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"DELETE\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/v1\/stars\/{star}\/face":{"post":{"tags":["Stars"],"operationId":"Update face's star","description":"```StarController::class```","requestBody":{"description":"```StarController::class```","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["image"],"properties":{"image":{"type":"image","example":"","description":"Face image of the star."}},"example":{"image":""}}}}},"parameters":[{"in":"path","name":"star","description":"ID of the star.","required":true,"schema":{"type":"integer","example":1}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Une image est requise.","errors":{"image":["Une image est requise."]}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\/face\" \n    -d \"image\"=\"\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\/face\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"image\": \"\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"image\" =>  \"\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/v1\/stars\/{star}\/face\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}}},"x-tagGroups":[{"name":"General","tags":["Stars"]}]}